WignerD.jl

Wigner D matrices computed using the phase convention of Varshalovich et al. (1988). In this conevntion, wavefucntions transform under rotation as

\[\psi_{jm}(\hat{n}) = \sum_{m^\prime=-j}^j D^{j*}_{m,m^\prime}(\alpha,\beta,\gamma) \psi_{jm^\prime}(\hat{n}^\prime),\]

where $\alpha$, $\beta$ and $\gamma$ are the Euler angles corresponding to the rotation. The Wigner D -matrix is related to the d-matrix through

\[D^{j}_{m,m^\prime}(\alpha,\beta,\gamma) = d^j_{m,m^\prime}(\beta) \exp(-i(m\alpha + m^\prime\gamma))\]

WignerD.Equator — Type

Equator

Alias for β = π/2 in spherical polar coordinates.

WignerD.NorthPole — Type

NorthPole

Alias for β = 0 in spherical polar coordinates.

WignerD.SouthPole — Type

SouthPole

Alias for β = π in spherical polar coordinates.

WignerD.wignerD — Function

wignerD(j, α::Real, β::Real, γ::Real, [Jy = zeros(ComplexF64, 2j+1, 2j+1)])

Evaluate the Wigner D-matrix with elements $D^j_{m,n}(α,β,γ)$ for the angular momentum $j$ and the Euler angles $α$, $β$ and $γ$. The momentum $j$ may be an integer or a half-integer, and must be non-negative. Optionally the pre-allocated matrix Jy may be provided, which must be a ComplexF64 matrix of size (2j+1, 2j+1), and may be overwritten during the calculation.

WignerD.wignerD! — Function

wignerD!(D, j, α::Real, β::Real, γ::Real, [Jy = zeros(ComplexF64, 2j+1, 2j+1)])

Evaluate the Wigner D-matrix with elements $D^j_{m,n}(α,β,γ)$ for the angular momentum $j$ and the Euler angles $α$, $β$ and $γ$, and store the result in D. The momentum $j$ may be an integer or a half-integer, and must be non-negative. Optionally the pre-allocated matrix Jy may be provided, which must be a ComplexF64 matrix of size (2j+1, 2j+1), and may be overwritten during the calculation.

WignerD.wignerDjmn — Method

wignerDjmn(j, m, n, α::Real, β::Real, γ::Real)

Evaluate the Wigner D-matrix element $D^j_{m,n}(\alpha,\beta,\gamma)$.

WignerD.wignerd — Function

wignerd(j, β::Real, [Jy = zeros(ComplexF64, 2j+1, 2j+1)])

Evaluate the Wigner d-matrix with elements $d^j_{m,n}(β)$ for the angular momentum $j$ and the rotation angle $β$. The momentum $j$ may be an integer or a half-integer, and must be non-negative. Optionally the pre-allocated matrix Jy may be provided, which must be a ComplexF64 matrix of size (2j+1, 2j+1), and may be overwritten during the calculation.

WignerD.wignerd! — Function

wignerd!(d, j, β::Real, [Jy = zeros(ComplexF64, 2j+1, 2j+1)])

Evaluate the Wigner d-matrix with elements $d^j_{m,n}(β)$ for the angular momentum $j$ and the rotation angle $β$, and store the result in d. The momentum $j$ may be an integer or a half-integer, and must be non-negative. Optionally the pre-allocated matrix Jy may be provided, which must be a ComplexF64 matrix of size (2j+1, 2j+1), and may be overwritten during the calculation.

WignerD.wignerdjmn — Method

wignerdjmn(j, m, n, β::Real)

Evaluate the Wigner d-matrix element $d^j_{m,n}(β)$.